Move code around in value
There are now 3 files in src/value/:

* mod.rs with basic code and trait impls
* ser.rs with Serialize and Serializer
* de.rs with Deserialize
